implemented PR reviews
